Atualização da estrutura inicial
