+++ src/endpoints/addEstudante.ts
@@ -1,0 +1,5 @@
+
+
+export const addEstudante = () => {
+    
+}
