+++ src/index.ts
@@ -1,25 +1,24 @@
-import express, { Express,Request, Response } from "express";
-import cors from "cors";
-import { AddressInfo } from "net";
-import { connection } from "./connection";
+import { config } from "dotenv"
+import app from "./app"
+import { addEstudante } from "./endpoints/addEstudante"
 
 
-const app: Express = express();
-app.use(express.json());
-app.use(cors());
+config()
+
+app.post('/users/signup', addEstudante)
 
 //Endpoint de teste
-app.get("/", async (req,res)=>{
-    const result = await connection.raw("SHOW TABLES")
-    console.log(result);
-    res.send("o endpoint está funcionando")
-})
+// app.get("/", async (req,res)=>{
+//     const result = await connection.raw("SHOW TABLES")
+//     console.log(result);
+//     res.send("o endpoint está funcionando")
+// })
 
-const server = app.listen(process.env.PORT || 3003, () => {
-    if (server) {
-       const address = server.address() as AddressInfo;
-       console.log(`Server is running in http://localhost: ${address.port}`);
-    } else {
-       console.error(`Failure upon starting server.`);
-    }
-});
+// const server = app.listen(process.env.PORT || 3003, () => {
+//     if (server) {
+//        const address = server.address() as AddressInfo;
+//        console.log(`Server is running in http://localhost: ${address.port}`);
+//     } else {
+//        console.error(`Failure upon starting server.`);
+//     }
+// });
